Removed Aegis Update, made the output time accurate pre horn
Removed the complication of editing the aegis expire time, hard to use well because it could only be used if this user picked up aegis and it would hold up the clipboard update too much and it was unreliable because aegis can be not picked up at all as well. The string outputs were wrong for pre-horn as 0:01 ingame is 0 seconds for pregame. So made curerntTime -1. Simplified some stuff and removed a few unneeded things.