Added short/long form option/output, close button, code to limit user spam,
Added a setting to allow user to select between short form and long form outputs depending on which they prefer. Added a close button so the user can just close without saving settings and not have to press more buttons to close. Added several things to stop a user from being able to spam the hotkey causing the GSL to not listen properly and to limit the possiblity of spam causing lag when it is trying to set data to the clipboard when the user is pasting at the same time. (match id check, last run currenttime, last run system time). Removed/cleaned up a few things that were not used. If the game is paused, the user will not be given an updated time when the new and old times are identical.  Updated download and manual cfg (sets items to 0 because there is no longer an aegis check)